+++ examples/swr-site/app/_components/video.tsx
@@ -2,7 +2,6 @@
 
 import { useCallback, useEffect, useRef } from 'react'
 import { useInView } from 'react-intersection-observer'
-import 'intersection-observer'
 
 export function Video({ src, caption, ratio }) {
   const [inViewRef, inView] = useInView({ threshold: 1 })