+++ packages/nextra-theme-docs/src/contexts/active-anchor.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import type { Dispatch, ReactElement, ReactNode, SetStateAction } from 'react'
-import 'intersection-observer'
 import { createContext, useContext, useRef, useState } from 'react'
 
 type ActiveAnchor = Record<