+++ packages/nextra/__test__/normalize-page.spec.ts
@@ -1,6 +1,6 @@
 import { normalizePages } from '../src/client/normalize-pages.js'
 import { normalizePageMap } from '../src/server/normalize-page-map.js'
-import { cnPageMap, usPageMap } from './fixture/page-maps/pageMap.js'
+import { cnPageMap, usPageMap } from './fixture/page-maps/page-map.js'
 
 describe('normalize-page', () => {
   it('zh-CN home', () => {