+++ packages/nextra-theme-blog/CHANGELOG.md
@@ -1,5 +1,19 @@
 # nextra-theme-blog
 
+## 3.0.0-alpha.36
+
+### Patch Changes
+
+- Updated dependencies [2b9b95b]
+  - nextra@3.0.0-alpha.36
+
+## 3.0.0-alpha.35
+
+### Patch Changes
+
+- Updated dependencies [f662237]
+  - nextra@3.0.0-alpha.35
+
 ## 3.0.0-alpha.34
 
 ### Patch Changes