+++ packages/nextra/CHANGELOG.md
@@ -1,5 +1,17 @@
 # nextra
 
+## 3.0.0-alpha.36
+
+### Patch Changes
+
+- 2b9b95b: migrate to `@headlessui/react` v2
+
+## 3.0.0-alpha.35
+
+### Patch Changes
+
+- f662237: avoid focus-visible style being cut off by overflow-hidden
+
 ## 3.0.0-alpha.34
 
 ### Patch Changes